Add custom S2I run code to complete database migration
The ASP.NET example app is deployed successfully with dotnet 1.0 image
but the register step is not working as 'no such table: AspNetUsers'
error will occur due to the fact the migrations for ApplicationDbContext
haven't been migrated to database CreateIdentitySchema yet. As a result,
the command 'dotnet ef database update' needs to be run to complete the
database migration process. In order to automatically run this command,
the custom S2I run code containing the command is needed. The S2I run
code is excuted when the application is run.

Signed-off-by: Vu Dinh <vdinh@redhat.com>